Run tests
This was done in the old AB::MB version, so we should do it here.
This was my intent all along, the requires for Test::More and
Text::Template at the top of the share block is to make sure that
our Perl has the necessary modules to run the test suite.